fix: address code review feedback for PR #38
- Remove hardcoded DEVELOPMENT_TEAM from project.pbxproj; reference
  $(DEVELOPMENT_TEAM) so each contributor supplies their own Team ID
  via the gitignored Config.xcconfig
- Add Config.xcconfig.example with DEVELOPMENT_TEAM placeholder
- Extract ActivityPresenter into Views/Helpers/ActivityPresenter.swift
  so PromptPresetsView and SettingsView share one implementation;
  also migrates SettingsView away from .sheet() to avoid black screen
- Add iPad popoverPresentationController configuration in
  ActivityPresenter to prevent crash when sourceView is missing
- Handle makeExportFile() returning nil with an error alert instead
  of silently leaving showShareSheet stuck in an inconsistent state
